pyproject.toml
[tool.poetry]
name = "lead-orchestrator"
version = "0.1.0"
description = ""
authors = ["Michael <mpgg2000@gmail.com>"]
readme = "README.md"
packages = [{include = "lead_orchestrator"}]

[tool.poetry.scripts]
server = "lead_orchestrator.main:run"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.89.0"
uvicorn = {extras = ["standart"], version = "^0.20.0"}
gunicorn = "^20.1.0"
multidict = "^6.0.4"
frozenlist = "^1.3.3"
python-dotenv = "^0.21.0"
redis = "^4.4.2"
amocrm-api-client = {git = "https://github.com/kiortir/amocrm_api_client.git"}
httpx = "^0.23.3"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
autopep8 = "^2.0.1"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
types-redis = "^4.4.0.0"

[build-system]
requires = ["poetry==1.2.2"]
build-backend = "poetry.core.masonry.api"
